Some minor corrections

+++ src/operators/test/testInverseJacobianBug.cc
@@ -72,7 +72,7 @@ void myTest(AMP::UnitTest *ut, std::string exeName) {
     std::vector<AMP::Mesh::MeshElement> nodes = el->getElements(AMP::Mesh::Vertex);
     for(size_t i = 0 ; i < nodes.size() ; i++) {
       std::vector<double> pt = nodes[i].coord();
-      fprintf(fp, "nd = %d, x = %.15lf, y = %.15lf, z = %.15lf \n", i, pt[0], pt[1], pt[2] );
+      fprintf(fp, "nd = %d, x = %.15lf, y = %.15lf, z = %.15lf \n", (int) i, pt[0], pt[1], pt[2] );
     }
   }
